[kernel][crypto]: Add EntropyPool.
EntropyPool represents a collection of entropy from user defined
sources. Its purpose is to represent an opaque container for the raw
types used internally in the PRNG impl, and will be used to seed the
PRNG with 256 bits of entropy.

EntropyPool is a movable and copyable  object, that scrapes itself once its been
moved through mandatory_memset. The copyable is required for snapshoting
on concurrent contexts for calls to Add.

By doing so, we can guarantee that the Prng is always initialized with
enough entropy bits, and we can clean it up(remove the non-thread-safe
thread safe bits of its interface.

Test: k ut crypto-entropy-pool
Change-Id: Ib939c13a40bf64c0a47c3e33d6252e9df4ef4102
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562149
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>