[cm_fidl_analyzer] Don't support dynamic offers.
Updating a TODO added in a previous CL which suggested that
cm_fidl_analyzer will need to support dynamic offers. In fact,
cm_fidl_analyzer is always running in a static context, so it's fine
to panic if it encounters a dynamic offer. This CL removes the TODO
and gives the panic! a clearer message.

Change-Id: I11ee8c9bc9bb4a0fa40b1a9b177bb4b84565ae4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582191
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>