[ftl] Remove inspect test and return integration test to CQ
Removes a test that broke as a result of Inspect behaviour. The
integration package was disabled from CQ as a result. This test mostly
just verified that the inspect nodes it creates at start up get created,
which is mostly a test of the Inspect library. So this test has been
removed and the package is being returned to CQ.

Since this was disabled we've had log severity restrictions added, so
this is disabled for this test as well since we actually want failure
to locate an NDM as an error log, even though it is normal for a new
device.


MULTIPLY: ftl-integration (fuchsia):20

Bug: 84931
Change-Id: I591fded6e6937d2c853e787155a63efed884b969
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583504
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Suter <csuter@google.com>