[kernel][ktrace] Refactor for testing.
Refactor the way that things like syscall names, probe names, thread
names, and so on, get added to the trace buffer immediately following
a rewind or start operation to facilitate testing.  Specifically; move
these operations into protected virtual methods which can be
overridden by a test version of the KTraceState class, so that the
result of the operations does not depend on the current state of the
system.

Change-Id: I1262f9e4281c5a1c75694bdf63b7fa3d5302bdb5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579744
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>