[build] Simplify clang_toolchain() template.
Refactor the Rust and C++ tool() definitions to de-duplicate most
of the computations (about how to build their command line), by using
a common scope, for each language, that contains most of the relevant
information.

For the things that need to vary depending on the tool() / output type,
use string templates that are expanded inside the tool() definition
instead.

This is similar to what zircon_toolchain() does, and will allow for
a better understanding and overall simplification of clang_toolchain()
and zircon_toolchain() in the future.

For testing, the content of out/default/ has been compared before/after
the CL was applied, the only differences found are in
unknown_wasm32/toolchain.ninja and are:

- The unknown_wasm32_{bin,dylib} tool commands now include an
  LD_LIBRARY_PATH="../../prebuilt/third_party/rust/linux-x64/lib"
  to help them find the right LLVM library version, just like
  the unknown_wasm32_cdylib one.

  This also requires removing the "-C link-args=-Wl,--Map..."
  argument from their command line, because the rust-lld tool
  does not seem to support them. Given that map files for
  wasm binaries are probably not useful to anyone, this is
  not a real loss for the build.

Apart from that, everything else is identical.

Bug: 74100

Change-Id: I0ba02f9774e28745bd44fcec6461b2e02c1d1f67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548466
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Shai Barack <shayba@google.com>