[fx format-code] Fail if unable to choose formatter
If a file doesn't have an extension or it has an unsupported extension,
`fx format-code` currently exits successfully without doing anything.
That's confusing because it makes it seem like the file is already
correctly formatted, when in fact fx doesn't actually know how to format
the file. A more user-friendly approach is to fail noisily in this
situation so the user knows that formatting failed, and why it failed.

Change-Id: I5b527f906d95177b8f4e400f49a36dd35dd0dafd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584061
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>