[kernel][crypto]: Rename PRNG->Prng.
All caps are usually reserved for macros. So its clear it is class for a
reader, when we follow this sort of pattern.
As a side not, GlobalPRNG is a namespace, so its been renamed to
global_prng to keep the same behavior. This is an uncommon pattern,
usually singletons are provided in the form of Class::Get/GetInstance.

Test: CQ, k ut all
Change-Id: Icc1a3097169b1b32dfe98071dedf6c717613e585
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562762
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>