[fx] Introduce new version of `fx sync-to`
This version is open-source and supports syncing to checkouts used by
public builds and tryjobs, unlike the legacy internal version, by
reading the `checkout_info` property emitted by a build to dynamically
determine the revision to checkout and which repos to patch. On the
other hand, the legacy version of `fx sync-to` downloads a jiri snapshot
from GCS, which doesn't work for syncing to a tryjob's checkout, and the
fact that it's internal means it's impossible to use it for reproducing
public builds' checkouts.

I implemented it in Python to make it easier to read and maintain, and
because it would have been extraordinarily painful to implement the
`checkout_info` JSON parsing logic in Bash in the legacy version of `fx
sync-to`.

I named the new version `new-sync-to` instead of just `sync-to` for now
so I have some time to test it before replacing the old implementation.
I also haven't yet implemented the logic to download a Jiri snapshot
from GCS, which is necessary to support builds that run on old release
branches from before recipes emitted the `checkout_info` property
(anything prior to f5).

I tested every possible "state" argument form on Linux, and only the
build ID form on Mac since jiri update is horrendously slow on my
laptop. Mac/Linux differences should be minimal compared to Bash scripts
anyway since Python is actually a reasonable language that's consistent
across all platforms.

Bug: 68492
Bug: 72533
Change-Id: Ib6d7a2fe121df15cd788e96b8ae5c7dcbf7b9ccf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582883
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>