[scenic] Lower log level
The "new error message" is collateral fallout
from a Flatland-related refactoring in Scenic
(fxr/570848), and the refactoring was tracked in
fxb/82643.

This message is only seen in Root Presenter, which
is a tightly controlled component, and "we know
what we're doing".

Bug: 83903
Change-Id: Ia2b526299eb64a3a47a86a84447abae1f61e3702
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584172
Fuchsia-Auto-Submit: Jaeheon Yi <jaeheon@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>