[cts][session] Remove cts_session_config
The cts_session_config rule globally configures the startup session
for all instances of session_manager in a build. This will cause
session_manager to crash on startup if it can't resolve and start
the session.

For example, the session_manager in core will crash if
reference_session_config is included in the build because it references
a relative component that doesn't exist in session_manager's package.

In the future, tests should either instruct session_manager to start
a session with the `fuchsia.session` APIs or explicitly pass
configuration to it, and not rely on global configuration.

Bug: 67088
Bug: 83905
Change-Id: I029845a0c9edd5010d7abf18f506fe338b7e56eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584244
Reviewed-by: John Shamoon <johnshamoon@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>