[cts][fidl] Add test for error handlers set after peer closed
Add a test that verifies that a binding error handler set after
the peer closes the channel does not get called.

Add a test that verifies that binding a request and setting an error
handler after a peer closes the channel calls the error handler.

Test: fx test //sdk/cts/tests/pkg/fidl/cpp
Change-Id: I5cc7efcc8c050ccddfec128316c0d64728435319
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579591
Commit-Queue: Ben Lawson <benlawson@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>