[kernel][vm] Update access information for inactive pages
When pages in PageQueuePagerBackedInactive are accessed, their stashed
queue number in the vm_page_t is updated, but since this queue is not
processed as part of ProcessLruQueue, the pages are never moved out of
the inactive queue. This means that we might need to fix up the queue
when peeking pages for eviction. Without this fix we might have been
evicting pages that were in the inactive queue at some point but had
since been accessed.

Fixes an existing bug in eviction from the inactive queue, not directly
related to hinting, but required for hinting to work as intended.

Test: kernel-unittest

Bug: 65334
Change-Id: I63872c3616032a03280feb0fbe41510a90631b42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584163
Fuchsia-Auto-Submit: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adrian Danis <adanis@google.com>