[llcpp] Switch to using v2 as the in-memory LLCPP format
This CL changes LLCPP envelopes (see envelope.h) to use v2 as the in
memory wire format. For the sake of reducing the size of the CL:
- The transformer is used to continue to encode v1 bytes on the wire
- The LLCPP API is unchanged

The idea is to change the in-memory format now and then gradually change
each of the pieces that interacts with it in the future.

Bug: 79578
Bug: 84116

Change-Id: If3ed11e4b2ce4e7ed72eea75da51ab6562c2a413
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574261
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Pascal-Louis Perez <pascallouis@google.com>