[testrunner] Fix nil tester check.
In go, when checking if an interface is nil, it could either be the nil
interface or a non-nil interface with a nil value. Checking tester==nil
checks the first case, but when we set `tester, err =
newFuchsiaSSHTester()`, even if newFuchsiaSSHTester() returns nil as the
tester, tester gets set to a non-nil interface with a nil value. Thus
the tester==nil check doesn't work and we need to check the underlying
value. We can fix this by having the newFuchsiaSSHTester(),
newFuchsiaSerialTester(), and newSubprocessTester() funcs return the
tester interface instead of a pointer to their respective testers. That
way when we return nil from those functions, it doesn't set the pointer
value to nil but properly sets the tester interface value to nil so the
`tester==nil` check works.

Bug: 85051
Change-Id: I1c5d4069edd1c57e01976cecdc7d4bf29ed257aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584145
Commit-Queue: Ina Huh <ihuh@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>