[kernel][object] Remove resource usage from MsiAllocation
MsiAllocations are currently creating resource objects to track the
blocks of vectors they allocate. However, this isn't actually suitable
due to the difference between GSIs and Vectors on x64, or with the idea
of what a vector actually is. In this case, a vector is managed by the
interrupt controller in question (GIC or IoApic) and the resource is not
used for any sort of ownership exclusivity or transfer. With this in
mind, the platform level defines what an Interrupt Resource corresponds
to and MSI will leave the management of those objects to the platform
rather than risking conflating the vector space with something else.

Change-Id: Icb009879276bc333b9f8815cf036ef2e27c56aee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575777
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>