[phys][mmu]: PageTableEntry alignment.
Force alignment to be 8 bytes. This was tripping with atomic operations
complaining about unaligned operations. Since the expected alignment was
8 but the actual alignment was 4.

No functional changes. The x86 code seems to enforce 8 byte alignment as
well. And the comment on the getters and setters suggest the same.

Test: build and CQ
Change-Id: If927d22546dc1ccb94683b83b8df1d3a4f10d61c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583194
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>