[hermetic] Allow license tool to be non-hermetic
The license checker scans the whole tree and so touches files outside of
its inputs. This change allows the license_data action to pass the
action tracer check, but it's not an ideal solution. In the future we
should change the way the notice files are generated so that it plays
better with incremental building. See fxb/84924 for some discussion.

Bug: 84924

Change-Id: I1dd3c63cd7b02bc938583feaea7eb1a1a7de25ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583027
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>