[a11y] Fall back to value when range_value is not present for sliders.
Some sliders don't have the range_value field set, but do supply the
range value in the value field. When the range_value is not set, we
should fall back to the value field to ensure that the user still gets
some feedback on the value.

Bug: 84613
TEST: fx test a11y-lib-tests, manual
Change-Id: Ibf79f5f23ea4556208665404c220d424434b16fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582775
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Alexander Brusher <abrusher@google.com>