[gidl][llcpp] Support new setters for LLCPP inlined envelopes
Iac59dea6aced870360e4624a6d83711e217477a2 added new setters for tables
and unions that take only a single parameter -- the object being set.
In an upcoming CL, the older setters will be removed when the inline
size is <= 4 bytes. Because of this, it is necessary to update the GIDL
tests to support the new setters.

Bug: 79578

Change-Id: Iaffe91976f43885c10618c781bbde11ec7bc3b36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583901
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>