[build][restat] Substitute when following a flag
Add the ability to condition a substitution on the previous option,
which is often a flag like '-o'.
This enables substitution of more specific occurrences of an output
name, and avoids some unintended substitutions.

This is needed particularly when the output file name happens to appear
in multiple places on the command-line, but only one instance is
intended for substitution.  For Rust tests and some binaries, the output
file name is also the crate name, and we don't intend to change the
crate name.  With Rust commands, the output file always follows `-o`.

Example: only substitute the option argument of `-x`:
  $this_script --outputs substitute_after:-x:bar.out --
  	command bar.out -x bar.out     bar.out -y bar.out
  will run:
  	command bar.out -x bar.out.tmp bar.out -y bar.out

This syntax was chosen to avoid a token starting with '-' being
misinterpreted by argparse as an option.

The behavior of `--outputs FILENAME` is the same as before

Updated use of --outputs in clang_toolchain.gni.

Bug: 85133
Change-Id: I86e7069ca10deec543740344abcdb53c040cc48f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584502
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: David Turner <digit@google.com>