[fx] Fix format-code warning logic
fxrev.dev/584061 made `fx format-code` fail and print an error if it
encountered a file it didn't know how to format.

This is useful for notification purposes if the user specifies a single
file that `fx format-code` can't format, but it's bad for `fx
format-code --git`, which tries to format all changed files. In that
case it's better to silently skip files that we don't know how to
format.

Likewise, the `--files` flag can take a glob, so it's possible that an
unformattable file is only included incidentally by a glob. Therefore,
we only show the error if the user specified a single file (or a glob
that only matched one file).

Change-Id: Ic0501af452fbfbc0ff37062f1e56ce4cfc0a326e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584344
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>