[cm_fidl_analyzer][routing] De/serialize AnalyzerModelError
In order to plumb errors from the routing static analyzer
through to Scrutiny, we'll need to be able to de/serialize
AnalyzerModelError and the various error types that it wraps.
The main intermediate step is to derive serde::{Deserialize,
Serialize} for RoutingError on host builds, replacing
'static borrows of error params with copies in a few places
to simplify lifetime management.

Bug: 61861

Change-Id: Ia15717edd6a121693272b9a10ec9dd438e069b02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584603
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Commit-Queue: Laura Peskin <pesk@google.com>