[audio][sherlock] Fix bitmask ordering
mpuryear@ found that we are incorrectly setting the bitmask used by
the SetActiveChannels API (used by the audio-driver-ctl -a parameter).
We invert ring_buffer_channels_to_use_bitmask such that:
The least significant bit corresponds to channel index 0
This per RingBuffer::SetActiveChannels() API specification, such that
the bitmask sequence matches the channel ordering in the data
ring-buffer itself.

This is not a change in the driver itself, but rather how we
configure the mapping in the board driver. Also we only change
the bitmask that is used for SetActiveChannel(), others are not
reversed in this CL.

The "/tmp/n" files below only have audio in the n channel.

Fixed: 85229

Test: audio-driver-ctl play /tmp/2 plays in Sherlock's woofer
Test: audio-driver-ctl play /tmp/3 plays in Sherlock's left tweeter
Test: audio-driver-ctl play /tmp/4 plays in Sherlock's right tweeter
Test: audio-driver-ctl -c 4 -a 2 plays Sherlock's woofer
Test: audio-driver-ctl -c 4 -a 4 plays Sherlock's left tweeter
Test: audio-drvier-ctl -c 4 -a 8 plays Sherlock's right tweeter
Test: All speakers play audio in normal usage.

Change-Id: I2f1d640963c65aa50c54a0bfede4572616c8b5f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584681
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>