[fshost] Inspect data for minfs rewrite progress
Adds progress indicator inspect nodes (which are exported at
bootstrap/fshost:root/minfs_upgrade:*) to let us monitor the progress of
the in-place minfs rewrite.

This is wired up to the fxfs rewrite for now to have it exercised
somewhere.  We should change it to trigger on the minfs rewrite once
that is ready.

Bug: 85437
Bug: 84973
Change-Id: I834bcb65bfa95d22f5242ca6b3395f648fd7b09e