[audio_core] temporary 4->2 channel-mapper
As a short-term device-specific workaround, the SourceReader class
that provides channel mapping from 4-channel to 2-channel has been
changed to simply copy channels 0 and 1 as-is, ignoring channels
2 and 3. Previous behavior was to emit the average of input channels
0 and 2 as output channel 0, and the average of input channels 1 and
3 as output channel 1.

We expect this workaround to be removed once the driver is updated for
the device in question. This is tracked with fxb/85201.

fx test audio_mixer_unittests

Bug: fxb/85201

Change-Id: Ibab12f58b3d9efb0321d8907e02ee3e482b51e40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584501
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>