[wlan][policy] addition of roaming framework
This adds structure for roaming to happen. When there is an ongoing
connection, we will get periodic connection data and then use that data
to decide whether to scan for a network to switch to. Only one scan
should be requested at a time. After this CL, a scan should never
happen; a future CL will add the logic to decide whether to do a scan.

Bug: 83744
Test: Added a unit test:
	fx test wlancfg-tests
Change-Id: I67f06b4be48569288ab7ce62c35482f68c0f781a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579055
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Kiet Tran <kiettran@google.com>
Reviewed-by: Hayden Nix <haydennix@google.com>