[kernel][vm] AlwaysNeed implies latency sensitive
Introduces a flag to address spaces and VMOs to indicate they are
latency sensitive, and that certain kinds of background reclamation
should be avoided on them. At the moment these kinds of reclamation are
page table eviction (for the address space) and zero page deduping (for
the VMO).

In this commit the latency sensitive flag is inferred as a result of
performing an AlwaysNeed hint against an address space, and the flag
automatically propagates to any VMO mapped into the address space, both
at that point in time and in the future.

This flag, and the way it is inferred and propagated, is a short term
solution and is expected to be changed in the future (see related bug).

Bug: 85056
Change-Id: I82e5f15b395cadedb057ada88a09de67ad244c12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579122
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>