[cm] `add_child` cleanup.
Moves the Discover action registration out of `add_child_internal`,
since that was the last thing that function does, and that seems
clearer than having a boolean indicate whether to do it or not.

Also adds some more context to the comments for `add_child` and
`add_child_for_test`.

Change-Id: Ia3a8735da40e1671ecde0a58ff5738ab6d476c4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584904
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>