[ffx] Implement `--recreate` for `ffx component run`
If a component created using `ffx component run` still exists in
the topology, the `--recreate` flag will destroy it and recreate
it. This is an alternative to the `--name` flag, allowing devs
to reuse the same component instance name. It also removes the
need for developers to call `ffx component destroy` manually.

Test: fx test ffx_component_run_lib_test
Change-Id: Ibea6d34c9d5c842c8c3080b3166ec124f10abef7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584901
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>
Reviewed-by: Andy Weiss <dragonbear@google.com>