[netcfg] Remove devfs_path arg
The path to devfs no longer needs to be configurable because it is the
same in netemul test realms as it is in production.

Also remove the netcfg-netemul component which was the sole remaining
usage of the argument, which is now unused.

Bug: 78610
Test: fx test netcfg-tests
Change-Id: Icd381e7e45cfa1df031502f3f247effc4afb9727
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584906
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tony Gong <gongt@google.com>