[sampler] Convert Integer -> MEMORY_USAGE in 1.0
Cobalt 1.0 didn't have a raw INTEGER type, but its closest fit was
MEMORY_USAGE, which is used in practice by current Cobalt clients.

Also fixed fvm.json and sysmem.json which happened to work due to
Integer being converted to EVENT_COUNT (which the underlying cobalt
metrics are defined as).

Bug: 85437
Bug: 84626
Change-Id: I22c602a1fcf6b0321128ed031478ac6195e9b0da