[paver] Additional logging for failed paves.
Currently the paver appears to be giving up due to premature file
ending. This information will confirm that by showing that the metadata
expects more left in the extent and yet the end of file is seen. Also
adds to the PayloadStreamer to show if it exits with leftover data.

Bug: 83638

Change-Id: I3742df7a59f38175b60cef5327bc7224b60b3265
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585161
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>