[fidl_c] Remove envelope implementation for FIDL C encode
This is in response to a TODO to perform validation of envelopes for
FIDL C bindings encoding. Envelopes aren't supported for FIDL C bindings
so doing so doesn't make sense.

Instead, this CL updates the FIDL C encode implementation to error
whenever types with envelopes are provided as input.

Fixed: 85067

Change-Id: I6c8901901f0d83fb6e8b17c7ef2d9823ae3b44de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583845
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>