[stress_tests] Stress Tests as TRF runner
This CL proposes a new test framework runner for
stress tests.

This runner abstracts away the test loop and allows
the actors to be written as components in multiple
languages.

In this model, a stress test is a component realm
with the following topology:

                 Test Root
		 /        \
       Test Subject       Actors

Actors are components that are created dynamically
by the runner in a collection. The test writer will
route capabilities from the test subject to the actor
collection. The actors will use these capabilities to
stress the test subject.

The actors are controlled by the runner using a
FIDL protocol: fuchsia.stresstest.Actor.

Actors offer a set of actions to the runner. The
runner coordinates between the different actors,
making each one run different actions in parallel.

The Test Runner Framework does not see this
internal implementation. Stress tests have a single
test case exposed to TRF that will error out when any
of the actors fail.

Multiply: stress-runner-integration-test
Test: fx test stress-runner-integration-test
Change-Id: Ic4af6e2deed6e3631e2829b30c8e811acb999317
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567861
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
API-Review: Christopher Johnson <crjohns@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Reviewed-by: Robert Ma <robertma@google.com>