[usb-cdc-ecm] Reduce "ignored interrupt" log message severity.
This can be triggered under some circumstances for some adapters. In
particular, some rt8153 adapters seem to start sending these spurious
interrupts when the device is unexpectedly rebooted but the adapter
remains powered (as is the case when using an externally-powered hub).
These interrupts don't seem to impact the functionality of the device
but the error message makes logs and the serial console virtually useless.

Change-Id: I2f10c3badfc2efc6453245ffa484bd6daaf262b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585903
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>