[hypervisor] Rename zx_vcpu_resume to zx_vcpu_enter
This change does the following:
1. Renames zx_vcpu_resume to zx_vcpu_enter
2. Renames the kernel internals to match the change in name
3. Updates the associated tests to match the new naming
4. Updates documentation and comments to match the new naming

For the kernel internals, the choice was made to remain consistent all
the way through, as we have in the past for the implementation of the
hypervisor.

This is the first part of the implementation of RFC-0128.

Change-Id: Ic13fb0abc21d0210d488aedf9df88f3da2b0d76b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585904
API-Review: Abdulla Kamar <abdulla@google.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>