[ermine] Removing loading indicator upon view close.
If a view is taking a while to load, we show the loading indicator. If
the user close the view using keyboard shortcut before it finishes
loading, the shell gets stuck in an invalid state, thinking the loading
is not over.

This change resets the flag if the view is closed while it is still
loading.

Fixed: 85143
Change-Id: I69fad2f0a78e6c7473471110a5fb3b4002bea431
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/585538
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>