[fidl_c] Make fidl_validate an internal function
fidl_validate is currently part of the public FIDL API, however it is
intended to be used only by HLCPP internally to validate messages. This
CL renames the function to make it clear that it shouldn't be depended
on.

Unfortunately, there are two usages of fidl_validate that were left in
because there is not currently any alternative public HLCPP API.
- One validates HLCPP types for a test and could be transitioned to
encoding with Encoder::NO_HEADER with wire format specified, but that
also uses an internal API.
- The other validates event bodies and doesn't have a mechanism to
access the type

Bug: 63427
Change-Id: I885ae4dc2b0e7516afaff99ec278e7445f8bf6fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585443
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
API-Review: Ian McKellar <ianloic@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>