[ffx] Heuristic to find SDK root.
If we're working out-of-tree and the sdk.root config setting is unset,
we can try to find an SDK manifest as a parent of the `ffx` tool's
directory and use that directory as the SDK root. This will obviate
the need for most users to set this config setting.

Fixed: 84342
Change-Id: Idef9076fcb0c0682c9a93cd43651c4abecca6df2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586489
Reviewed-by: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Jordon Wing <jwing@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>