[bt][hfp] Peer Connected events are sent after connection completes
Before this, spurious connection results would be sent if a Peer
connects but does not make an L2CAP/RFCOMM connection for the
HFP Service Level Connection.

Also sets a timeout on an RFCOMM connection from the peer.
If the timeout is reached, the peer task is stopped.

Test: Added unit tests
Bug: 82575

Change-Id: Ib1911470e6e1b4335b9c21515a425ce942649488
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585623
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Matthew Kehrt <kehrt@google.com>