[FMS] Use a mapping file for product bundle paths
The user would have to pass in a list of product bundle paths
via a GN arg which was painful. This makes it so the user
passes a mapping file location that contains the name of the product
and the path to the product bundle for that product.

Bug: b/198166472
Change-Id: Ic15340e2ae744140878a54fee6708853c8f6c5e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585490
Commit-Queue: Omer Lev-ran <omerlevran@google.com>
Reviewed-by: Misha Gridnev <gridman@google.com>