[fserve] Add opt-in support for ffx repository
This starts the transition of fserve to using the new package repository
server that is built in ffx. In order to ease the transition, this is
made opt-in so that it can be tested out in field to shake out any
issues before being switched to the default.

Fixed: 84290

Change-Id: Ib8b290b96b95325f57026a3e0fa3b9e38f3b1b5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583629
Reviewed-by: Omer Lev-ran <omerlevran@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>