[ffx][assembly] Don't hardcode number of minfs partitions in fvm creation.
This is part of a change to support a minfs "account" partition in FVM
that will be encrypted with a key derived from the user's
authentication factor. This is in addition to the existing minfs
"data" partition that is encrpyted with a per-device key.

Prior to this CL the fvm logic in ffx was discarding all but the last
minfs partition in the board_config.

Bug: 582846
Change-Id: I5a6a22317e7f94ea938d4148069d59bc879a3ac6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586504
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>