[run-test-suite] Make name optional when verifying artifacts
The names of files are not guaranteed to be stable, although the
current implementation of the directory output uses static names for
each file. With custom artifacts, multiple components in the same test
may produce a custom artifact, so using a static name won't be
sufficient. To allow testing in this case, this change updates the
testing module to key artifacts by the artifact metadata instead of
filename. This allows verifying the artifacts while allowing
implementation flexibility in choosing filenames.

Change-Id: I158b9e67335e35d6d23201a551c8d74653b81e07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583624
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>