[component][fuzzing] Add Runner class.
This CL introduces the runner class, which is responsible for performing
the details of various fuzzing workflows. It uses a dedicated thread for
these workflows to avoid blocking the FIDL dispatcher thread. It owns
and manages the corpora and the mutagen to generate inputs, and the
module pool to collect and analyze feedback.

Bug: 24740
Change-Id: Ibbb0d0704c1a122740a429d2dac7c29ef0c87fdc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567108
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>